+++ src/internal/nse-defuse.c
@@ -6,11 +6,9 @@ sexp* rlang_capturearginfo(sexp* call, sexp* op, sexp* args, sexp* rho);
 sexp* rlang_capturedots(sexp* call, sexp* op, sexp* args, sexp* rho);
 
 sexp* rlang_ext2_capturearginfo(sexp* call, sexp* op, sexp* args, sexp* env) {
-  args = eval_ext2_args(args, env);
-  return rlang_capturearginfo(call, op, args, env);
+  return rlang_capturearginfo(call, op, r_node_cdr(args), env);
 }
 
 sexp* rlang_ext2_capturedots(sexp* call, sexp* op, sexp* args, sexp* env) {
-  args = eval_ext2_args(args, env);
-  return rlang_capturedots(call, op, args, env);
+  return rlang_capturedots(call, op, r_node_cdr(args), env);
 }