+++ src/internal/utils.h
@@ -9,12 +9,11 @@ sexp* rlang_enquo(sexp* sym, sexp* frame);
 
 extern sexp* rlang_ns_env;
 
- __attribute__((noreturn))
+__attribute__((noreturn))
+static inline
 void never_reached(const char* fn) {
   r_abort("Internal error in `%s()`: Reached the unreachable.", fn);
 }
-
-sexp* eval_ext2_args(sexp* args, sexp* env);
 
 
 #endif